created a test file for mongodb dataBase testing
