+++ commands/ping.js
@@ -3,6 +3,5 @@ module.exports = {
   description: 'Ping!',
   execute(msg, args) {
     msg.reply('pong');
-    msg.channel.send('pong');
   },
 };
