+++ commands/register.js
@@ -1,0 +1,9 @@
+module.exports = {
+  name: '!register',
+  description: 'Register User',
+  execute(msg, args) {
+    msg.reply('Enter User name');
+    
+
+  }
+};
